requirements.txt
alembic==1.1.0
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
mysqlclient==1.4.4
protobuf==3.9.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.16.0
WTForms==2.2.1
